% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_distance.R
\name{auc_distance}
\alias{auc_distance}
\title{Calculate the AUC Distance between Two DataFrames}
\usage{
auc_distance(df1, df2)
}
\arguments{
\item{df1}{A data frame representing the first dataset.}

\item{df2}{A data frame representing the second dataset.}
}
\value{
A numeric value representing the AUC (Area Under the ROC Curve) for the classification model.
}
\description{
This function calculates the Area Under the Curve (AUC) for a classification model trained on two datasets (df1 and df2).
It combines both datasets, assigns a class label to each, and then trains a Naive Bayes classifier to distinguish between
the two datasets. The AUC is computed by evaluating the model's performance on a random sample from the combined data.
}
\details{
The function combines df1 and df2 by adding a new column \code{class} with values \code{df1} for the first dataset
and \code{df2} for the second dataset. A Naive Bayes model is trained on the combined data to predict the \code{class} column.
A random sample of 1000 observations is then selected from the combined data and used to calculate the AUC score
based on the predicted probabilities.

The AUC value represents the classifier's ability to distinguish between the two datasets, where:
\itemize{
\item A higher AUC value (closer to 1) indicates better performance.
\item An AUC value of 0.5 indicates no discriminatory power (equivalent to random guessing).
}
}
\keyword{internal}
