% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_back_original.R
\name{convert_back_original}
\alias{convert_back_original}
\title{Convert Synthetic Uniform Data Back to Original Domain}
\usage{
convert_back_original(data_model, synthetic_uniform)
}
\arguments{
\item{data_model}{A data model generated by create_model}

\item{synthetic_uniform}{A data frame where each column represents a synthetic variable
sampled from a uniform distribution in \link{0,1}.}
}
\value{
A data frame with synthetic data transformed back to its original scale.
}
\description{
This function transforms synthetic data from the uniform \link{0,1} space back to
the original data domain using inverse ECDF for categorical variables and quantiles
for continuous variables.
}
\examples{
synthetic_uniform <- data.frame(A = runif(1000), B = runif(1000))
data_type=list(A='numeric',B='numeric')
df_model <- create_model(df, data_type)
converted_data <- convert_back_original(df_model,synthetic_uniform)
}
\keyword{internal}
