% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data_types.R
\name{convert_data_types}
\alias{convert_data_types}
\title{Convert Data Types in a Data Frame}
\usage{
convert_data_types(df, df_struct)
}
\arguments{
\item{df}{A data frame whose columns data type need to be converted}

\item{df_struct}{A reference data frame structure containing column names (\code{col_names})
and their corresponding data types (\code{col_type}).}
}
\value{
A data frame with standardized column data types.
}
\description{
This function ensures that the columns in a given data frame (\code{df}) are converted
to the appropriate data types based on a reference structure (\code{df_struct}).
}
\examples{
df <- data.frame(
  category = c("A", "B", "A", "C"),
  timestamp = c("01/01/2023 12:00:00", "02/01/2023 14:30:00",
                "03/01/2023 09:45:00", "04/01/2023 16:20:00")
)
df_struct <- data.frame(
  col_names = c("category", "timestamp"),
  col_type = c("factor", "datetime")
)
df_converted <- convert_data_types (df, df_struct)

}
\keyword{internal}
