% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{create_metadata}
\alias{create_metadata}
\title{Create a Data Frame Structure Definition}
\usage{
create_metadata(df, overrides = list())
}
\arguments{
\item{df}{A data frame for which the structure is to be created.}

\item{overrides}{A named list of column types to override default detections (optional).}
}
\value{
A data frame with two columns: \code{col_names} (column names) and \code{col_type} (detected or overridden data types).
}
\description{
This function generates a structure definition for a given data frame (\code{df}).
It automatically detects column types and allows for manual overrides.
}
\examples{
ed_struct <- create_metadata(ed, 
  overrides = list(
    establishment_code = "factor",
    arrival_datetime = "datetime",
    sex = "factor",
    ethnicity = "factor",
    triage_category = "factor",
    mode_of_arrival = "factor",
    presenting_complaint = "factor",
    admission = "factor",
    attendance_length_of_episode = "integer"
  )
)
}
\keyword{internal}
