% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{.onLoad}
\alias{.onLoad}
\title{Internal List of Distance Functions}
\usage{
.onLoad(libname, pkgname)
}
\description{
This list contains various distance functions used in the package.
It is intended for internal use.
}
\keyword{internal}
