% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_bootstrap.R
\name{ensemble_bootstrap}
\alias{ensemble_bootstrap}
\title{Ensemble Bootstrap Function}
\usage{
ensemble_bootstrap(
  data_model,
  df1,
  df2,
  iter = 1000,
  ncores = 1,
  distance = "auc"
)
}
\arguments{
\item{data_model}{A data model object containing metadata and original data.}

\item{df1}{A data frame representing the first candidate dataset.}

\item{df2}{A data frame representing the second candidate dataset.}

\item{iter}{An integer specifying the number of iterations for the Monte Carlo simulation (default is 1000).}

\item{ncores}{An integer specifying the number of cpu core if running parallel}

\item{distance}{A string specifying the distance metric to be used for comparison (default is 'auc').}
}
\value{
A data frame representing the best ensemble dataset after iterating through the Monte Carlo simulation.
}
\description{
This function performs a Monte Carlo simulation to generate an ensemble dataset by sampling rows from \code{df1} and \code{df2}. It computes a selected distance measure (e.g., AUC) between a model's data (\code{df_n}) and the candidate datasets (\code{df1} and \code{df2}), selects the best ensemble dataset based on minimizing the distance, and returns the best ensemble.
}
\examples{
ensemble_result <- ensemble_bootstrap(data_model, df1, df2, iter = 1000, ncore=1, distance = 'auc')
}
