% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_ga.R
\name{ensemble_ga}
\alias{ensemble_ga}
\title{Generate an Ensemble using Genetic Algorithm}
\usage{
ensemble_ga(data_model, df1, df2, GENERATIONS = 1000, distance = "frobenius")
}
\arguments{
\item{data_model}{A data model that includes conversion functions and dataset handling.}

\item{df1}{A data frame, the first source dataset.}

\item{df2}{A data frame, the second source dataset.}

\item{distance}{A string specifying the distance metric to use. Default is "auc".}
}
\value{
A data frame representing the best synthetic dataset generated through the genetic algorithm.
}
\description{
This function uses a genetic algorithm to create a synthetic dataset by combining
two input datasets (\code{df1} and \code{df2}) based on a specified distance metric. It employs
crossover, mutation, and selection processes to evolve a population of candidate solutions.
The final ensemble is selected based on the fitness of the individuals in the population.
}
\details{
The algorithm operates in multiple generations, each consisting of the following steps:
\itemize{
\item Fitness evaluation of the population.
\item Selection of the top half of the population.
\item Crossover and mutation to create new individuals for the next generation.
\item Elitism: The best individual is preserved across generations.
}
The fitness of each individual is determined by a distance function (such as AUC).
The final ensemble is derived from the individual with the best fitness score.
}
\examples{
\dontrun{
# Example usage
result <- ensemble_ga(data_model, df1, df2, distance = 'auc')
head(result)
}

}
