% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frobenius_distance.R
\name{frobenius_distance}
\alias{frobenius_distance}
\title{Compute Frobenius Distance Between Two DataFrames}
\usage{
frobenius_distance(df1, df2)
}
\arguments{
\item{df1}{A numeric data frame representing the first dataset.}

\item{df2}{A numeric data frame representing the second dataset.}
}
\value{
A numeric value representing the \strong{Frobenius distance} between the correlation matrices of the two datasets.
}
\description{
This function calculates the \strong{Frobenius distance} between the correlation matrices of two datasets (\code{df1} and \code{df2}). The Frobenius distance is a measure of the difference between two matrices, computed as the norm of the difference between their correlation matrices.
}
\details{
The function performs the following steps:
\enumerate{
\item \strong{Compute the correlation matrices} for both \code{df1} and \code{df2} using pairwise complete observations.
\item \strong{Calculate the Frobenius distance}: The Frobenius distance between the two correlation matrices is computed as the \strong{Frobenius norm} of their difference, which is the square root of the sum of the absolute squares of the elements in the difference matrix.
}

The Frobenius distance is commonly used to measure how similar or different two matrices are. In this case, it is used to measure the difference between the correlation structures of the two datasets.
}
