% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_syn_copula.R
\name{generate_syn_copula}
\alias{generate_syn_copula}
\title{generate_syn_copula Function: Generates synthetic data using Gaussian copula}
\usage{
generate_syn_copula(data_model, n_samples = NA)
}
\arguments{
\item{data_model}{A list containing the original dataset (\code{original}), transformed uniform dataset (\code{data_uniform}), and metadata (\code{metadata}).
The metadata should include column names (\code{col_names}).}

\item{n_samples}{The number of synthetic samples to generate. If not provided, the function defaults to the number of rows in the original dataset.}
}
\value{
A data frame containing synthetic data with the same structure as the input dataset. The data will have the same column names as the original dataset,
and the correlation structure will match the original dataset's correlation structure.
}
\description{
This function generates synthetic data by modeling the correlation structure of the input dataset using a Gaussian copula.
It estimates the copula parameters from the original dataset and uses them to generate new synthetic samples.
The generated synthetic data maintains the same correlation structure as the original dataset.
}
\details{
\itemize{
\item The function computes the correlation matrix of the uniform-transformed data and uses it to create a Gaussian copula.
\item The copula is fitted to the dataset, and new synthetic samples are generated by sampling from the fitted copula.
\item The synthetic data is returned in a uniform space, preserving the correlation structure of the original dataset.
}
}
\examples{
# Assuming data_model is a list containing the original dataset and metadata
}
\seealso{
\code{\link{create_model}}
synthetic_data <- generate_syn_copula(data_model, n_samples = 1000)
}
