% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_syn_decomp.R
\name{generate_syn_decomp}
\alias{generate_syn_decomp}
\title{Generate Synthetic Data Using Decomposition Methods}
\usage{
generate_syn_decomp(data_model, n_samples = NA)
}
\arguments{
\item{data_model}{A list containing the original data, transformed uniform data,
and metadata. The uniform data should be in \link{0,1} scale.}

\item{n_samples}{Integer. Number of synthetic samples to generate. If NA, defaults
to the number of rows in \code{data_model$original}.}
}
\value{
A data frame containing synthetic data in \link{0,1} uniform scale.
}
\description{
This function generates synthetic data from a given data model using either
Cholesky or eigen decomposition, depending on numerical stability conditions.
}
\details{
The function follows these steps:
\enumerate{
\item Converts input data to a normal space using the quantile function.
\item Computes the correlation matrix of the transformed data.
\item Ensures the correlation matrix is positive definite using \code{nearPD} if necessary.
\item Selects the decomposition method:
\itemize{
\item Eigen decomposition is used if the correlation matrix has a high condition
number, is large, close to singular, or still not positive definite.
\item Otherwise, Cholesky decomposition is used.
}
\item Generates synthetic data using random noise and the chosen decomposition method.
\item Transforms synthetic data back to a uniform scale using the normal CDF.
}
}
\note{
This function applies \code{nearPD} to adjust the correlation matrix if it is not
positive definite, ensuring numerical stability. If eigen decomposition is used,
non-negative eigenvalues are enforced.
}
\examples{
\dontrun{
  synthetic_data <- generate_syn_decomp(data_model)
}

}
