% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_number.R
\name{to_number}
\alias{to_number}
\title{to_number Function: Convert factors and datetime columns to numeric}
\usage{
to_number(data_model, df)
}
\arguments{
\item{data_model}{A list containing the metadata of the dataset, including column names
and their corresponding types. The metadata should be in the form of a dataframe with
at least two columns: 'col_names' (character vector of column names) and 'col_type'
(character vector of column types).}

\item{df}{A dataframe where columns will be converted based on the column types specified
in \code{data_model$metadata}.}
}
\value{
A modified dataframe with 'factor' columns converted to numeric and 'datetime' columns
converted to POSIX numeric timestamps.
}
\description{
This function takes a dataframe and converts columns of type 'factor'
to numeric values and columns of type 'datetime' to POSIX numeric timestamps.
It uses the \code{data_model} metadata to identify the column types.
}
\details{
\itemize{
\item \strong{Factors}: Converted to numeric values using \code{as.factor} and \code{as.numeric}.
\item \strong{Datetime}: Converted to POSIX timestamps using \code{as.POSIXct} and then to numeric.
}
}
\examples{
# Assuming `data_model` and `df` are predefined:
df_modified <- to_number(data_model, df)

}
\keyword{internal}
