% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasserstein_distance_costm.R
\name{wasserstein_distance_costm}
\alias{wasserstein_distance_costm}
\title{Compute Wasserstein Distance between Two Datasets}
\usage{
wasserstein_distance_costm(df1, df2)
}
\arguments{
\item{df1}{A numeric data frame representing the first dataset.}

\item{df2}{A numeric data frame representing the second dataset.}
}
\value{
A numeric value representing the \strong{Wasserstein distance} between the two datasets.
}
\description{
This function calculates the \strong{Wasserstein distance} (or \strong{Earth Mover's Distance}) between two datasets (\code{df1} and \code{df2}) based on their pairwise Euclidean distances. The Wasserstein distance is a measure of the "cost" required to transform one distribution into another. This function is commonly used to compare two datasets in terms of their distributions.
}
\details{
The function calculates the \strong{Wasserstein distance} between the two datasets by:
\enumerate{
\item \strong{Converting the datasets to matrices} (\code{df1} and \code{df2}) for compatibility with distance functions.
\item \strong{Computing the pairwise cost matrix} using the \strong{Euclidean distance} between all pairs of observations from \code{df1} and \code{df2}.
\item Using the \strong{Wasserstein} transport method from the \code{transport} package, the function calculates the optimal "flow" needed to transform the distribution of \code{df1} into the distribution of \code{df2} based on the pairwise distances.
}

The \strong{Wasserstein distance} is a useful metric in many applications like comparing distributions, generating synthetic data, and evaluating how well one dataset approximates another.
}
\keyword{internal}
