% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasserstein_distance_pp.R
\name{wasserstein_distance_pp}
\alias{wasserstein_distance_pp}
\title{Compute Wasserstein Distance using Point Patterns}
\usage{
wasserstein_distance_pp(df1, df2)
}
\arguments{
\item{df1}{A numeric data frame representing the first dataset.}

\item{df2}{A numeric data frame representing the second dataset.}
}
\value{
A numeric value representing the \strong{Wasserstein distance} between the point patterns of the two datasets.
}
\description{
This function calculates the \strong{Wasserstein distance} between two datasets (\code{df1} and \code{df2}) using \strong{point patterns}. The function converts the data into point patterns and computes the Wasserstein distance between their respective distributions. The Wasserstein distance measures the minimal "cost" of transforming one point pattern distribution into another.
}
\details{
The function calculates the \strong{Wasserstein distance} by:
\enumerate{
\item \strong{Resampling the datasets}: If one dataset has more rows than the other, the larger dataset is randomly resampled (without replacement) to match the number of rows in the smaller dataset.
\item \strong{Converting data to point patterns}: The function uses the \code{pp()} function to convert the datasets (\code{df1} and \code{df2}) into point patterns. This process involves treating the data as spatial points in a coordinate space.
\item \strong{Computing the Wasserstein distance}: The \strong{Wasserstein distance} is then calculated between the two point patterns using the \code{wasserstein()} function, which measures the minimal cost required to transform one point pattern into another.
}

The \strong{Wasserstein distance} is commonly used in spatial statistics and topological data analysis to compare point distributions or spatial patterns.
}
\keyword{internal}
